feat: complete modules cjsToEsm
