+++ src/fs/copy.js
@@ -1,7 +1,4 @@
-// implement function that copies folder files files with all its content
-// into folder files_copy at the same level (if files folder doesn't
-// exists or files_copy has already been created Error with message FS
-// operation failed must be thrown)
+// rewrite it to it's equivalent in ECMAScript notation (and rename it to esm.mjs)
 
 import { dirname, join } from "path";
 import { fileURLToPath } from "url";
